feat: Refactor application structure and integrate Tailwind CSS
- Removed existing CSS styles and replaced with Tailwind CSS integration.
- Updated App component to use ClipboardBody instead of Windows.
- Created new Clipboard component to manage clipboard history.
- Added Footer and Header components for improved layout.
- Implemented History component to display clipboard items with various types.
- Configured TypeScript paths for easier imports.
- Updated Vite configuration to include Tailwind CSS plugin and path aliasing.

+++ eslint.config.js
@@ -49,7 +49,7 @@ export default [
       'no-trailing-spaces': 'warn',
       'eol-last': ['warn', 'always'],
       'semi': ['warn', 'always'],
-      'quotes': ['warn', 'single'],
+      'quotes': ['warn', 'double'],
       'indent': ['warn', 2, { SwitchCase: 1 }]
     },
     settings: {
